+++ packages/launchpad/types/index.d.ts
@@ -96,7 +96,7 @@ export {
 } from "./pubkey";
 export { findSequenceForSignedTx } from "./sequence";
 export { encodeSecp256k1Signature, decodeSignature } from "./signature";
-export { AccountData, Algo, PrehashType, OfflineSigner } from "./signer";
+export { AccountData, Algo, OfflineSigner } from "./signer";
 export { CosmosFeeTable, SigningCosmosClient } from "./signingcosmosclient";
 export { isStdTx, pubkeyType, CosmosSdkTx, PubKey, StdFee, StdSignature, StdTx } from "./types";
 export { makeCosmoshubPath, executeKdf, KdfConfiguration } from "./wallet";