+++ packages/launchpad/types/signer.d.ts
@@ -1,6 +1,5 @@
 import { StdSignDoc } from "./encoding";
 import { StdSignature } from "./types";
-export declare type PrehashType = "sha256" | "sha512" | null;
 export declare type Algo = "secp256k1" | "ed25519" | "sr25519";
 export interface AccountData {
   /** A printable address (typically bech32 encoded) */
@@ -19,9 +18,5 @@ export interface OfflineSigner {
    * @param signerAddress The address of the account that should sign the transaction
    * @param signDoc The content that should be signed
    */
-  readonly sign: (
-    signerAddress: string,
-    signDoc: StdSignDoc,
-    prehashType?: PrehashType,
-  ) => Promise<StdSignature>;
+  readonly sign: (signerAddress: string, signDoc: StdSignDoc) => Promise<StdSignature>;
 }